requirements.txt
fastapi
python-jose
pydantic
uvicorn